+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.67.1
 * Fixed race of stop internal processes on close topic writer
 * Fixed goroutines leak within topic reader on network problems